+++ javax/src/test/java/me/panpf/javax/test/util/StopwatchTest.java
@@ -1,7 +1,7 @@
 package me.panpf.javax.test.util;
 
-import me.panpf.javax.lang.Numberx;
 import me.panpf.javax.util.Collectionx;
+import me.panpf.javax.util.Rangex;
 import me.panpf.javax.util.Stopwatch;
 import org.junit.Assert;
 import org.junit.Test;
@@ -19,7 +19,7 @@ public final class StopwatchTest {
         Thread.sleep(1000L);
 
         Stopwatch stopwatch = new Stopwatch();
-        Assert.assertTrue(Numberx.in(stopwatch.startTime, time + 1000L, time + 1010L));
+        Assert.assertTrue(Rangex.in(stopwatch.startTime, time + 1000L, time + 1010L));
     }
 
     @Test
@@ -32,21 +32,21 @@ public final class StopwatchTest {
 
         Thread.sleep(1000L);
         Stopwatch.CountLap firstCountLap = stopwatch.countLap();
-        Assert.assertTrue(Numberx.in(firstCountLap.time, stopwatch.startTime + 1000L, stopwatch.startTime + 1010L));
-        Assert.assertTrue(Numberx.in(firstCountLap.distanceLastTime, 1000L, 1010L));
-        Assert.assertTrue(Numberx.in(firstCountLap.distanceStartTime, 1000L, 1010L));
+        Assert.assertTrue(Rangex.in(firstCountLap.time, stopwatch.startTime + 1000L, stopwatch.startTime + 1010L));
+        Assert.assertTrue(Rangex.in(firstCountLap.distanceLastTime, 1000L, 1010L));
+        Assert.assertTrue(Rangex.in(firstCountLap.distanceStartTime, 1000L, 1010L));
 
         Thread.sleep(2000L);
         Stopwatch.CountLap secondCountLap = stopwatch.countLap();
-        Assert.assertTrue(Numberx.in(secondCountLap.time, stopwatch.startTime + 3000L, stopwatch.startTime + 3020L));
-        Assert.assertTrue(Numberx.in(secondCountLap.distanceLastTime, 2000L, 2020L));
-        Assert.assertTrue(Numberx.in(secondCountLap.distanceStartTime, 3000L, 3020L));
+        Assert.assertTrue(Rangex.in(secondCountLap.time, stopwatch.startTime + 3000L, stopwatch.startTime + 3020L));
+        Assert.assertTrue(Rangex.in(secondCountLap.distanceLastTime, 2000L, 2020L));
+        Assert.assertTrue(Rangex.in(secondCountLap.distanceStartTime, 3000L, 3020L));
 
         Thread.sleep(3000L);
         Stopwatch.CountLap thirdCountLap = stopwatch.countLap();
-        Assert.assertTrue(Numberx.in(thirdCountLap.time, stopwatch.startTime + 6000L, stopwatch.startTime + 6030L));
-        Assert.assertTrue(Numberx.in(thirdCountLap.distanceLastTime, 3000L, 3030L));
-        Assert.assertTrue(Numberx.in(thirdCountLap.distanceStartTime, 6000L, 6030L));
+        Assert.assertTrue(Rangex.in(thirdCountLap.time, stopwatch.startTime + 6000L, stopwatch.startTime + 6030L));
+        Assert.assertTrue(Rangex.in(thirdCountLap.distanceLastTime, 3000L, 3030L));
+        Assert.assertTrue(Rangex.in(thirdCountLap.distanceStartTime, 6000L, 6030L));
     }
 
     @Test